+++ packages/vm/src/calls.rs
@@ -3,7 +3,7 @@ use serde::de::DeserializeOwned;
 use std::fmt;
 use wasmer::Val;
 
-use cosmwasm_std::{ContractResult, Env, MessageInfo, QueryResponse, Response, SubCallResult};
+use cosmwasm_std::{ContractResult, Env, MessageInfo, QueryResponse, Reply, Response};
 
 use crate::backend::{BackendApi, Querier, Storage};
 use crate::conversion::ref_to_u32;
@@ -90,10 +90,10 @@ where
     Ok(result)
 }
 
-pub fn call_subcall_response<A, S, Q, U>(
+pub fn call_reply<A, S, Q, U>(
     instance: &mut Instance<A, S, Q>,
     env: &Env,
-    msg: &SubCallResult,
+    msg: &Reply,
 ) -> VmResult<ContractResult<Response<U>>>
 where
     A: BackendApi + 'static,
@@ -103,7 +103,7 @@ where
 {
     let env = to_vec(env)?;
     let msg = to_vec(msg)?;
-    let data = call_subcall_response_raw(instance, &env, &msg)?;
+    let data = call_reply_raw(instance, &env, &msg)?;
     let result: ContractResult<Response<U>> = from_slice(&data)?;
     Ok(result)
 }
@@ -197,9 +197,9 @@ where
     call_raw(instance, "system", &[env, msg], MAX_LENGTH_SYSTEM)
 }
 
-/// Calls Wasm export "subcall_response" and returns raw data from the contract.
+/// Calls Wasm export "reply" and returns raw data from the contract.
 /// The result is length limited to prevent abuse but otherwise unchecked.
-pub fn call_subcall_response_raw<A, S, Q>(
+pub fn call_reply_raw<A, S, Q>(
     instance: &mut Instance<A, S, Q>,
     env: &[u8],
     msg: &[u8],
@@ -210,12 +210,7 @@ where
     Q: Querier + 'static,
 {
     instance.set_storage_readonly(false);
-    call_raw(
-        instance,
-        "subcall_response",
-        &[env, msg],
-        MAX_LENGTH_SUBCALL_RESPONSE,
-    )
+    call_raw(instance, "reply", &[env, msg], MAX_LENGTH_SUBCALL_RESPONSE)
 }
 
 /// Calls Wasm export "query" and returns raw data from the contract.